tested all .py files - they work

+++ example_01_imdb_binary_classification.py
@@ -25,7 +25,7 @@ from keras.datasets import imdb
 
 
 def multi_hot_encode(sequences, num_classes):
-    results = np.zeros((len(sequences), num_classes))
+    results = np.zeros((len(sequences), num_classes), dtype="float32")
     for i, sequence in enumerate(sequences):
         results[i][sequence] = 1.0
     return results
